VgaUtil: Lowercase jpg filename in disp_image_file().
This fixes the display of some images like losegame.jpg. The game was
passing in an uppercase filename to disp_image_file() in some places,
and thus failed to find the requested images.

Lowercasing the filename here, as opposed to doing it at the call sites,
avoids repetition and minimizes the number of changes needed.