ImageRes: Add new overloaded variant of put_to_buf().
This is required to properly display the info panel in other resolutions
besides 800x600 and 1024x768. Without this, the info panel will be
placed at the wrong position and be overlaid by the map.

The new function takes two additional ints, x and y, that specify the
position of the info panel. The original function simply calls the new
one with 0 for both x and y. This allows the callers of the original
function to continue working without modification.