Campaign: Fix crash during attack animation on strategic map.
A heap buffer overflow in random_unused_race() caused an invalid Fryhtan
race id to be returned. As a result, attackerArray was initialized with
bogus values, which, in turn, led to a failed attempt to initialize
attackerUnit in attack_animation(CampaignAnimationUnit...) and crashed
the game.

Using MAX_MONSTER_TYPE instead of MAX_RACE when a random Fryhtan race is
requested ensures that the race id lies within a valid range.

This bug was tracked down with the help of AddressSanitizer.